Removed all mentions of metrics2
