--- server/manager/src/test/resources/hadoop-metrics2-accumulo.properties
@@ -1,60 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-#
-
-# Poll collectors every thirty seconds
-*.period=5
-
-#
-# MetricSink configuration
-#
-# Uncomment a sink (and configuration) to enable it. By default,
-# no sinks are enabled.
-#
-
-#
-# Configure file logging
-#
-
-# File sink for all metrics
-accumulo.sink.file-all.class=org.apache.hadoop.metrics2.sink.FileSink
-accumulo.sink.file-all.filename=./target/it.all.metrics
-
-accumulo.sink.file-gc.class=org.apache.hadoop.metrics2.sink.FileSink
-accumulo.sink.file-gc.context=accgc
-accumulo.sink.file-gc.filename=./target/accgc.metrics
-accumulo.sink.file-gc.period=5
-
-# File sink for tserver metrics
-# accumulo.sink.file-tserver.class=org.apache.hadoop.metrics2.sink.FileSink
-# accumulo.sink.file-tserver.context=tserver
-# accumulo.sink.file-tserver.filename=tserver.metrics
-
-# File sink for manager metrics
-accumulo.sink.file-manager.class=org.apache.hadoop.metrics2.sink.FileSink
-accumulo.sink.file-manager.context=manager
-accumulo.sink.file-manager.filename=./target/manager.metrics
-accumulo.sink.file-manager.period=5
-
-accumulo.jmx.manager.context=manager
-
-# File sink for thrift server metrics
-# accumulo.sink.file-thrift.class=org.apache.hadoop.metrics2.sink.FileSink
-# accumulo.sink.file-thrift.context=thrift
-# accumulo.sink.file-thrift.filename=thrift.metrics
-
