+++ server/tserver/pom.xml
@@ -59,7 +59,6 @@
     <dependency>
       <groupId>io.micrometer</groupId>
       <artifactId>micrometer-core</artifactId>
-      <version>${micrometer.metrics.version}</version>
     </dependency>
     <dependency>
       <groupId>org.apache.accumulo</groupId>
