+++ test/src/main/java/org/apache/accumulo/test/metrics/TestStatsDSink.java
@@ -24,8 +24,10 @@ import java.net.DatagramPacket;
 import java.net.DatagramSocket;
 import java.net.SocketException;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
-import java.util.concurrent.ConcurrentLinkedQueue;
+import java.util.Map;
+import java.util.concurrent.LinkedBlockingQueue;
 
 import org.apache.accumulo.core.util.threads.Threads;
 import org.slf4j.Logger;
@@ -33,10 +35,64 @@ import org.slf4j.LoggerFactory;
 
 public class TestStatsDSink implements Closeable {
 
+  public static class Metric {
+    private final String name;
+    private final String value;
+    private final String type;
+    private final Map<String,String> tags = new HashMap<>();
+
+    public Metric(String name, String value, String type) {
+      super();
+      this.name = name;
+      this.value = value;
+      this.type = type;
+    }
+
+    public String getName() {
+      return name;
+    }
+
+    public String getValue() {
+      return value;
+    }
+
+    public String getType() {
+      return type;
+    }
+
+    public Map<String,String> getTags() {
+      return tags;
+    }
+
+    @Override
+    public String toString() {
+      return "Metric [name=" + name + ", value=" + value + ", type=" + type + ", tags=" + tags
+          + "]";
+    }
+  }
+
+  public static Metric parseStatsDMetric(String line) {
+    int idx = line.indexOf(':');
+    String name = line.substring(0, idx);
+    int idx2 = line.indexOf('|');
+    String value = line.substring(idx + 1, idx2);
+    int idx3 = line.indexOf('|', idx2 + 1);
+    String type = line.substring(idx2 + 1, idx3);
+    int idx4 = line.indexOf('#');
+    String tags = line.substring(idx4 + 1);
+    Metric m = new Metric(name, value, type);
+    String[] tag = tags.split(",");
+    for (String t : tag) {
+      String[] p = t.split(":");
+      m.getTags().put(p[0], p[1]);
+    }
+    return m;
+  }
+
   private static final Logger LOG = LoggerFactory.getLogger(TestStatsDSink.class);
 
   private final DatagramSocket sock;
-  private final ConcurrentLinkedQueue<String> received = new ConcurrentLinkedQueue<>();
+  private final LinkedBlockingQueue<String> received = new LinkedBlockingQueue<>();
 
   public TestStatsDSink() throws SocketException {
     sock = new DatagramSocket();
@@ -68,7 +124,9 @@ public class TestStatsDSink implements Closeable {
   }
 
   public List<String> getLines() {
-    return new ArrayList<String>(received);
+    List<String> metrics = new ArrayList<>(received.size());
+    received.drainTo(metrics);
+    return metrics;
   }
 
   public String getLine() {
