--- test/src/main/resources/accumulo.metrics.configuration.properties
@@ -1,33 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-#
-
-metrics.service.01.name=LoggingRegistry
-metrics.service.01.classname=org.apache.accumulo.server.metrics.service.file.LoggingMetricsRegistration
-metrics.service.01.enabled=true
-
-metrics.service.02.name=PrometheusRegistry
-metrics.service.02.classname=org.apache.accumulo.server.metrics.service.prometheus.PrometheusMetricsRegistration
-metrics.service.02.enabled=true
-metrics.service.02.manager.PortNumber=9234
-metrics.service.02.gc.PortNumber=9235
-metrics.service.02.tserver.PortNumber=9236
-
-metrics.service.03.name=JmxRegistry
-metrics.service.03.classname=org.apache.accumulo.server.metrics.service.JmxRegistryLoader
-metrics.service.03.enabled=false
